Refactor Navbar.jsx to index.jsx
